feat: add clipboard history module with header, item rows, hooks, and utilities
- Implemented ClipboardHistoryHeader component for displaying title, count, filters, and actions.
- Created ClipboardHistoryItemRow component for individual clipboard history entries with editing capabilities.
- Developed hooks for managing clipboard history items, feedback, and selection.
- Added utility functions for clipboard text normalization, filtering, and history persistence.
- Defined types for clipboard history items and related components.
- Configured constants for clipboard history settings and limits.
