+++ src/renderer/components/components/StandaloneWindowViewport.tsx
@@ -31,7 +31,7 @@ import { UrlFavoritesTab } from '../states/maxExpand/components/UrlFavoritesTab'
 import { AlbumTab } from '../states/maxExpand/components/album/components/AlbumTab';
 import { MailTab } from '../states/maxExpand/components/MailTab';
 import { LocalFileSearchTab } from '../states/maxExpand/components/LocalFileSearchTab';
-import { ClipboardHistoryTab } from '../states/maxExpand/components/clipBoardHistory/components/ClipboardHistoryTab';
+import { ClipboardHistoryTab } from '../states/maxExpand/components/clipBoardHistory';
 import { SettingsTab } from '../states/maxExpand/components/SettingsTab';
 import { MemoTab } from '../states/maxExpand/components/MemoTab';
 import { AlarmTab } from '../states/maxExpand/components/alarm/components/AlarmTab';
