+++ src/renderer/components/states/maxExpand/MaxExpandContentEager.tsx
@@ -31,7 +31,7 @@ import { AiChatTab } from './components/agent/components/AiChatTab';
 import { TodoTab } from './components/TodoTab';
 import { UrlFavoritesTab } from './components/UrlFavoritesTab';
 import { LocalFileSearchTab } from './components/LocalFileSearchTab';
-import { ClipboardHistoryTab } from './components/clipBoardHistory/components/ClipboardHistoryTab';
+import { ClipboardHistoryTab } from './components/clipBoardHistory';
 import { AlbumTab } from './components/album/components/AlbumTab';
 import { MailTab } from './components/MailTab';
 import { SettingsTab } from './components/SettingsTab';
